% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_junc.R
\name{add_junc_pos}
\alias{add_junc_pos}
\title{adds junction position in transcripts}
\usage{
add_junc_pos(tx, jx)
}
\arguments{
\item{tx}{transcripts a \code{\link[GenomicRanges]{GRangesList}} with
transcripts defined as GRanges of exons.}

\item{jx}{splice junctions as GRanges objects}
}
\value{
an integer vector with junction positions
}
\description{
adds junction position in transcripts
}
