% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_peptide.R
\name{add_peptide}
\alias{add_peptide}
\title{Annotate splice junctions with resulting CDS and peptide sequence}
\usage{
add_peptide(df, size = 30, bsg = NULL)
}
\arguments{
\item{df}{A data.frame with splice junctions in rows and at least the colums:
\itemize{
\item \code{junc_id} junction id consisting of genomic coordinates
\item \code{tx_id} the ID of the affected transcript (see \code{\link{add_tx}})
\item \code{cds_lst} a list of \code{\link[GenomicRanges]{GRanges}} with the
CDS (see \code{\link{add_tx}})
}}

\item{size}{the total size of the output sequence (might be shorter if peptide is shorter)}

\item{bsg}{\code{\link[BSgenome]{BSgenome}} object such as
\code{\link[BSgenome.Hsapiens.UCSC.hg19]{BSgenome.Hsapiens.UCSC.hg19}}}
}
\value{
A data.frame as the input with the additional column(s):
\itemize{
\item \code{cds_alt_lst} a list of \code{\link[GenomicRanges]{GRanges}} with the
modified CDS ranges.
\item \code{cds_id_alt} an identifier made from \code{tx_id} and \code{junc_id}
\item \code{junc_pos_cds} the junction position in the modified CDS sequence
\item \code{protein} the full protein sequence of the translated modified CDS.
\item \code{protein_junc_pos} The position of the junctino in the \code{protein} sequence
\item \code{junc_in_orf} Indicator whehter the junction is located in an open reading frame.
\item \code{peptide_context_seq_raw} the peptide sequence around the junction includig stop codons.
\item \code{peptide_context} the peptide sequence around the junction truncated after stop codons.
\item \code{peptide_context_junc_pos} The junction position relative to the \code{peptide_context} sequence
}
}
\description{
Annotate splice junctions with resulting CDS and peptide sequence
}
\examples{
requireNamespace("BSgenome.Hsapiens.UCSC.hg19", quietly = TRUE)
bsg <- BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19

junc_df <- toy_junc_df \%>\%
  mutate(
    cds_lst = as.list(toy_cds[tx_id])
  )

add_peptide(junc_df, size = 30, bsg = bsg)

}
