% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_context_sequence.R
\name{add_transcript_coordinates}
\alias{add_transcript_coordinates}
\title{adds transcript coordinates in the resulting transcript to the genomic
ranges of the exome}
\usage{
add_transcript_coordinates(transcript_range)
}
\arguments{
\item{transcript_range}{Genomic range of a transcript}
}
\value{
The Genomic range of a transcript with transcript coordinates
}
\description{
adds transcript coordinates in the resulting transcript to the genomic
ranges of the exome
}
