% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_peptide.R
\name{annot_neo}
\alias{annot_neo}
\title{annotate peptide sequence with wt and novel parts}
\usage{
annot_neo(cds_mod, cds_wt, bsg)
}
\arguments{
\item{cds_mod}{a named \code{\link[GenomicRanges]{GRangesList}} of the
modified (mutated) coding sequences (CDS) ranges}

\item{cds_wt}{a named \code{\link[GenomicRanges]{GRangesList}} of the
wild-type (unmodifed)coding sequences (CDS) ranges}

\item{bsg}{\code{\link[BSgenome]{BSgenome}} object such as
\code{\link[BSgenome.Hsapiens.UCSC.hg19]{BSgenome.Hsapiens.UCSC.hg19}}}
}
\value{
A logical list indicating for each amino-acid residue of the modified
transcripts whether it is novel peptide or wild-type
}
\description{
annotate peptide sequence with wt and novel parts
}
