% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_tx.R
\name{annotate_exon_idx}
\alias{annotate_exon_idx}
\title{Annotate splice junctions- transcripts combination and transcripts with exon details}
\usage{
annotate_exon_idx(df)
}
\arguments{
\item{df}{A data.frame with junctions- transcripts combination in rows.
\itemize{
\item \code{junc_id} junction id consisting of genomic coordinates
\item \code{tx_id} junction id consisting of genomic coordinates
\item \code{tx_lst} a list of \code{\link[GenomicRanges]{GRanges}} with the transcript
}}
}
\value{
A data.frame as the input but with more annotations.
}
\description{
Annotate splice junctions- transcripts combination and transcripts with exon details
}
