% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmsplice.R
\name{annotate_mmsplice}
\alias{annotate_mmsplice}
\title{Annotates the mmsplice output with additional columns including the junction as junc_id.}
\usage{
annotate_mmsplice(mmsplice_df, transcripts)
}
\arguments{
\item{mmsplice_df}{A data.frame like object from mmsplice output. This should at least
have the following columns:
- ID
- exon_id
- exons
- transcript_id
- delta_logit_psi}

\item{transcripts}{a GRangesList with transcripts defined as GRanges of exons
created by \code{GenomicFeatures::exonsBy(txdb, by = c("tx"), use.names = TRUE)}.
The exons in individual GRanges are assumed to be sorted according to
transcription sense, that for transcript with positive strand by position and
by descending position for transcripts on negative strand.}
}
\value{
A data.frame like object like the input but with additional columns:
\itemize{
\item junc_id: \verb{<chr>_<pos1>_<pos2>_<strand>}
\item event_type: either \code{exon_skipping} or \code{exon_inclusion}
}

If the logit delta PSI score is <= 0, the event is treated as exon skipping.
In this case a junction is build from the end of the exon before, and the
start of the exon after that.

If the logit delta PSI score is > 0, the event is treated as cassette exon.
In this case a (canonical) junction is build from the end of the upstream
exon to the start of the affected exon and from the end of the affected exon
to the start of the next exon.

MMsplice predicts the change on percent spliced in (PSI) for a given
annotated exon. Therefore, only exon inclusion and exon skipping needs to be
converted to junctions.
\itemize{
\item For \emph{exon skipping} the end of the previous and start of the next exon
build the junction
\item For \emph{exon inclusion} the (canonical) junction is build
from the end of the upstream exon to the start of the affected exon and from
the end of the affected exon  to the start of the next exon.
}
}
\description{
Annotates the mmsplice output with additional columns including the junction as junc_id.
}
