% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_spliceai_junction.R
\name{annotate_spliceai_junction}
\alias{annotate_spliceai_junction}
\title{Annotate splice variants with resulting junctions}
\usage{
annotate_spliceai_junction(var_df, transcripts, transcripts_gr)
}
\arguments{
\item{var_df}{a data.frame wit variants and (at least) the following columns:
\itemize{
\item \code{CHROM}
\item \code{POS}
\item \code{ALT}
\item \code{ALT}
\item \code{change} an change effect class from \emph{SpliceAI}. One of \code{DL}, \code{DG}, \code{AL}, \code{AG}.
\item \code{pos_rel} affected position relative to \code{POS}.
}}

\item{transcripts}{a GRangesList with transcripts defined as GRanges of exons
created by \code{GenomicFeatures::exonsBy(txdb, by = c("tx"), use.names = TRUE)}.}

\item{transcripts_gr}{a GRanges object with transcript created by
\code{GenomicFeatures::transcripts(txdb)}}
}
\description{
Annotate splice variants with resulting junctions
}
\examples{
spliceai_file <- system.file("extdata", "spliceai_output.vcf", package = "splice2neo")
df_raw <- parse_spliceai(spliceai_file)
df <- format_spliceai(df_raw)

annotate_spliceai_junction(df, toy_transcripts, toy_transcripts_gr)


}
