% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_peptide.R
\name{annotate_truncated_cds}
\alias{annotate_truncated_cds}
\title{Tests if junction leads to truncated mutated gene product
\code{protein_wt}
\code{protein}
\code{protein_len}
\code{junc_in_orf}}
\usage{
annotate_truncated_cds(df)
}
\arguments{
\item{df}{tibble}
}
\description{
Tests if junction leads to truncated mutated gene product
\code{protein_wt}
\code{protein}
\code{protein_len}
\code{junc_in_orf}
}
\keyword{internal}
