% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_to_junc.R
\name{bed_to_junc}
\alias{bed_to_junc}
\title{Transform a bed file into junction format}
\usage{
bed_to_junc(bed_file, type = "exon-exon")
}
\arguments{
\item{bed_file}{Path to the bed file}

\item{type}{Use for this parameter
\itemize{
\item \code{exon-exon} if the bed file defines exon-exon boundaries
\item \code{intron} if the bed file defines introns
}}
}
\value{
A character vector with junction IDs
}
\description{
Transform a bed file into junction format
}
