% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_tx.R
\name{calc_distance_exitron}
\alias{calc_distance_exitron}
\title{Calculate distance of splice sites of an exitron to next canonical splice sites in the transcript.}
\usage{
calc_distance_exitron(pos, tx_lst, pos_exon_idx, site = "left")
}
\arguments{
\item{pos}{A \code{\link[GenomicRanges]{GRanges}} representation of the exitron splice junction.}

\item{tx_lst}{A list of \code{\link[GenomicRanges]{GRanges}} with the transcript.}

\item{pos_exon_idx}{Exon index in transcript if splice site overlaps exon.}

\item{site}{A string describing the site of the exitron to be annotated.}
}
\value{
An integer of the genomic distance in bp.
}
\description{
Calculate distance of splice sites of an exitron to next canonical splice sites in the transcript.
}
\keyword{internal}
