% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_tx.R
\name{calc_distance_regular}
\alias{calc_distance_regular}
\title{Calculate distance of splice site to next canonical splice site in the transcript.}
\usage{
calc_distance_regular(
  pos,
  tx_lst,
  pos_exon_idx,
  pos_exon_boundary,
  site = "left"
)
}
\arguments{
\item{pos}{A \code{\link[GenomicRanges]{GRanges}} representation of the splice junction.}

\item{tx_lst}{A list of \code{\link[GenomicRanges]{GRanges}} with the transcript.}

\item{pos_exon_idx}{Exon index in transcript if splice site overlaps exon.}

\item{pos_exon_boundary}{Exon index in transcript if junction matches exon boundary.}

\item{site}{A string describing the site of the junction to be annotated.}
}
\value{
An integer of the genomic distance in bp.
}
\description{
Calculate distance of splice site to next canonical splice site in the transcript.
}
\keyword{internal}
