% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{canonical_juncs}
\alias{canonical_juncs}
\title{A tibble with canonical junctions and their source (comma, separated).
example data}
\format{
A tibble frame with 2 variables
}
\source{
random
}
\usage{
canonical_juncs
}
\description{
A tibble with canonical junctions and their source (comma, separated).
example data
}
\keyword{datasets}
