% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gtf.R
\name{canonical_junctions}
\alias{canonical_junctions}
\title{Build canonical junctions from transcripts}
\usage{
canonical_junctions(tx)
}
\arguments{
\item{tx}{a \code{\link[GenomicRanges]{GRangesList}} of reference transcripts}
}
\value{
a character vector of canonical splice junction ids
}
\description{
Build canonical junctions from transcripts
}
