% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_spliceai_junction.R
\docType{data}
\name{change_to_junction_rules}
\alias{change_to_junction_rules}
\title{Rules on how a splicing affecting variant creates a junction}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 4 columns.
}
\usage{
change_to_junction_rules
}
\description{
Rules on how a splicing affecting variant creates a junction
}
\keyword{datasets}
