% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_tx.R
\name{choose_tx}
\alias{choose_tx}
\title{Select transcripts per junction that are more likely to be affected by a junction. All possible transcripts can be annotated with \code{add_tx()}}
\usage{
choose_tx(df)
}
\arguments{
\item{df}{A data.frame with splice junctions in rows and at least the columns:
\itemize{
\item \code{junc_id} junction id consisting of genomic coordinates
\item \code{tx_id} junction id consisting of genomic coordinates
\item \code{tx_lst} a list of \code{\link[GenomicRanges]{GRanges}} with the transcript
}}
}
\value{
A data.frame as with relevant transcript and junction combinations. If tx_id is NA in the input data.frame such rows are removed from the output data.frame.

NOTE: This function selects transcripts that are more likely to be affected to reduce data dimension.
It excludes transcripts in which both junction positions are located in an intron. Junctions in a given transcript must be either be classified as
ES, IR, exitron or ASS event or have both junction positions in an exon. Other junction-transcript combinations are also excluded.
This function may loose relevant or keep irrelevant junction-transcripts in particular in regions with mutliple isoforms with distinct splicing pattern.
}
\description{
Select transcripts per junction that are more likely to be affected by a junction. All possible transcripts can be annotated with \code{add_tx()}
}
