% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_tx.R
\name{classify_junc_tx}
\alias{classify_junc_tx}
\title{Classify splice junctions- transcripts combinations based on junction and coordinates of annotated exons}
\usage{
classify_junc_tx(df)
}
\arguments{
\item{df}{A data.frame with junctions- transcripts combination in rows.}
}
\value{
A data.frame as the input but with more annotations in additional columns.
}
\description{
Classify splice junctions- transcripts combinations based on junction and coordinates of annotated exons
}
