% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_tx.R
\name{classify_transcripts}
\alias{classify_transcripts}
\title{Classify splice junctions- transcripts combinations based on junction and coordinates of annotated exons}
\usage{
classify_transcripts(df)
}
\arguments{
\item{df}{A data.frame with unctions- transcripts combination in rows.
#'}
}
\value{
A data.frame as the input but with more annotations.

#'
}
\description{
Classify splice junctions- transcripts combinations based on junction and coordinates of annotated exons
}
