% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_spliceai_mmpslice.R
\name{combine_mut_junc}
\alias{combine_mut_junc}
\title{Combines tibbles with junctions from SpliceAI and MMsplice}
\usage{
combine_mut_junc(spliceai_juncs, mmsplice_juncs)
}
\arguments{
\item{spliceai_juncs}{A tibble the junctions identified by SpliceAI}

\item{mmsplice_juncs}{A tibble the junctions identified by MMsplice in
standardized format}
}
\value{
A combined table with unique junctions. The columns DNA_tool
contains information which tools identified the given junction
}
\description{
Combines tibbles with junctions from SpliceAI and MMsplice
}
\examples{

}
