% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutated_transcript_ranges.R
\name{construct_mutated_range}
\alias{construct_mutated_range}
\title{returns genomic range of mutated transcript according to the splice event}
\usage{
construct_mutated_range(
  junc_pos1,
  junc_pos2,
  wt_transcript_range,
  strand_direction
)
}
\arguments{
\item{junc_pos1}{GRanges object for junction position 1}

\item{junc_pos2}{GRanges object for junction position 2}

\item{wt_transcript_range}{Genomic range of wildtype transcript in which the
alternative splicing event is taking place}

\item{strand_direction}{strand direction. Shall be "+" or "-"}
}
\value{
A tibble with columns \code{junc_id}, \code{chr}, \code{pos1}, \code{pos2}, \code{strand},
\code{jidx}, \code{subjectHits}, \code{enst}. This tibble returns all transcripts that are
affected by a given junction.
}
\description{
returns genomic range of mutated transcript according to the splice event
}
