% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutated_transcript_ranges.R
\name{construct_range_both_in_exon}
\alias{construct_range_both_in_exon}
\title{returns genomic range of mutated transcript for junction with both coordinates
in an exon}
\usage{
construct_range_both_in_exon(
  exon1_index,
  exon2_index,
  transcript_range,
  strand_direction,
  junction_start,
  junction_end
)
}
\arguments{
\item{exon1_index}{index of exon related to first position of junction in
transcript ranges}

\item{exon2_index}{index of exon related to second position of junction in
transcript ranges}

\item{transcript_range}{Genomic range of mutated transcript in which
the alternative splicing event is taking place}

\item{strand_direction}{strand direction. Shall be "+" or "-"}

\item{junction_start}{junction start coordinate}

\item{junction_end}{junction end coordinate}
}
\value{
a mutated transcript range based on the presented parameters
}
\description{
returns genomic range of mutated transcript for junction with both coordinates
in an exon
}
