% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_tx.R
\name{distance_to_splice_site}
\alias{distance_to_splice_site}
\title{Annotate for each junction the distance to the next canoncial donor and acceptor splice sites.}
\usage{
distance_to_splice_site(df)
}
\arguments{
\item{df}{A data.frame with junctions- transcripts combination in rows.
\itemize{
\item \code{junc_id} junction id consisting of genomic coordinates
\item \code{tx_id} transcript id
\item \code{tx_lst} a list of \code{\link[GenomicRanges]{GRanges}} with the transcript
\item \code{pos_left_on_exon_end} exon index if junction start matches with exon end coordinate
\item \code{pos_right_on_exon_start}  exon index if junction end matches with exon start coordinate
\item \code{pos_left_on_exon_idx} exon index in transcript if junction start overlaps exon
\item \code{pos_right_on_exon_idx} exon index in transcript if junction end overlaps exon
}

and additional columns.}
}
\value{
A data.frame of the input annotated with distances to canoncial splice sites:
\itemize{
\item \code{distance_to_next_canonical_donor} Genomic distance to the next canonical donor in the transcript
\item \code{distance_to_next_canonical_acceptor} Genomic distance to the next canonical acceptor in the transcript
}
}
\description{
Annotate for each junction the distance to the next canoncial donor and acceptor splice sites.
}
\keyword{internal}
