% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_mut_effect.R
\docType{data}
\name{effect_to_junction_rules_wo_ir}
\alias{effect_to_junction_rules_wo_ir}
\title{Rules on how a splicing affecting variant creates a junction without intron
retention}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 4 columns.
}
\usage{
effect_to_junction_rules_wo_ir
}
\description{
Rules on how a splicing affecting variant creates a junction without intron
retention
}
\keyword{internal}
