% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_context_sequence.R
\name{extract_sequence_window}
\alias{extract_sequence_window}
\title{adds transcript coordinates in the resulting transcript to the genomic
ranges of the exome}
\usage{
extract_sequence_window(
  transcript_sequence,
  transcript_range,
  junc_pos1,
  window_size = 200
)
}
\arguments{
\item{transcript_sequence}{Sequence of the full transcript}

\item{transcript_range}{Genomic range of a transcript}

\item{junc_pos1}{Genomic range of the first junction coordinate}

\item{window_size}{number of nucleotides left and right from the "breakpoint"}
}
\value{
The Genomic range of a transcript with transcript coordinates
}
\description{
adds transcript coordinates in the resulting transcript to the genomic
ranges of the exome
}
