% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irfinder_parsing.R
\name{filter_irfinder_txt}
\alias{filter_irfinder_txt}
\title{Filter IRFinder intermediate table to remove likely
false postive IR predictions}
\usage{
filter_irfinder_txt(tib, warnings, ratio_cutoff)
}
\arguments{
\item{tib}{IRFinder intermediate tibble with columns Warnings and IRratio}
}
\value{
A tibble in standardized junction format
}
\description{
Filter IRFinder intermediate table to remove likely
false postive IR predictions
}
