% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_junc_id.R
\name{generate_junction_id}
\alias{generate_junction_id}
\title{Given the chromosome, junction start, junction end and strand,
a junction id is created that follows the format: \verb{<chr>:<start>-<end>:<strand>}}
\usage{
generate_junction_id(chr, start, end, strand)
}
\arguments{
\item{chr}{The chromosome}

\item{start}{The start of the junction}

\item{end}{The end of the junction}

\item{strand}{Strand information. "+" or "-"}
}
\value{
The junction id in the format \verb{<chr>:<start>-<end>:<strand>}
}
\description{
Given the chromosome, junction start, junction end and strand,
a junction id is created that follows the format: \verb{<chr>:<start>-<end>:<strand>}
}
\examples{
generate_junction_id(chr = "chr1", start = 50, end = 100, strand = "+")

}
