% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_context_sequence.R
\name{get_context_sequence}
\alias{get_context_sequence}
\title{Given the coordinates of junctions( junc1_gr, junc2_gr, strand.dir),
returns the context sequence}
\usage{
get_context_sequence(
  transcript_id,
  transcript_db,
  junc_pos1,
  junc_pos2,
  strand_direction,
  genome_db,
  window_size = 200
)
}
\arguments{
\item{transcript_id}{An ENSEMBL transcript id}

\item{transcript_db}{a GRangesList with transcripts composed of exon ranges}

\item{junc_pos1}{GRanges object for junction position 1}

\item{junc_pos2}{GRanges object for junction position 2}

\item{strand_direction}{strand direction. Shall be "+" or "-"}

\item{genome_db}{The genome as DNAStringSet of the genomic sequences of each chromosome or a BSGenome object e.g. BSgenome.Hsapiens.UCSC.hg19}

\item{window_size}{number of nucleotides left and right from the "breakpoint"}
}
\value{
A tibble with columns \code{junc_id}, \code{chr}, \code{pos1}, \code{pos2}, \code{strand},
\code{jidx}, \code{subjectHits}, \code{enst}. This tibble returns all transcripts that are
affected by a given junction.
}
\description{
Given the coordinates of junctions( junc1_gr, junc2_gr, strand.dir),
returns the context sequence
}
