% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_junc_pos.R
\name{get_intronretention_alt_pos}
\alias{get_intronretention_alt_pos}
\title{Get the alternative position of input junction from an intron retention event
in the transcript sequences}
\usage{
get_intronretention_alt_pos(tx_lst, tx_mod, jx, intron_retention)
}
\arguments{
\item{tx_lst}{transcripts a \code{\link[GenomicRanges]{GRangesList}} with
transcripts defined as GRanges of exons.}

\item{tx_mod}{modified transcripts a \code{\link[GenomicRanges]{GRangesList}} with
transcripts defined as GRanges of exons.}

\item{jx}{splice junctions as GRanges objects}

\item{intron_retention}{logical vector indicating if junctions belong to an intron retention event}
}
\value{
an integer vector with alternative junction positions of intron
retention event
}
\description{
Get the alternative position of input junction from an intron retention event
in the transcript sequences
}
