% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_peptide.R
\name{get_normalized_protein_junc_pos}
\alias{get_normalized_protein_junc_pos}
\title{Annotate the normalized junction position in the protein,
\code{frame_shift}
\code{is_intron_retention}
\code{protein_junc_pos}
\code{protein}
\code{protein_wt}
\code{junc_pos_cds}
\code{is_first_reading_frame}
\code{junc_pos_cds_wt}
\code{cds_length_difference}}
\usage{
get_normalized_protein_junc_pos(df)
}
\arguments{
\item{df}{tibble}
}
\description{
Annotate the normalized junction position in the protein,
\code{frame_shift}
\code{is_intron_retention}
\code{protein_junc_pos}
\code{protein}
\code{protein_wt}
\code{junc_pos_cds}
\code{is_first_reading_frame}
\code{junc_pos_cds_wt}
\code{cds_length_difference}
}
\keyword{internal}
