% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_peptide.R
\name{get_peptide_context}
\alias{get_peptide_context}
\title{Get peptide context sequence given a tibble with
\code{normalized_protein_junc_pos}
\code{is_first_reading_frame}
\code{frame_shift}
\code{additional_novel_AA}
\code{protein}
\code{cds_length_difference}}
\usage{
get_peptide_context(df, flanking_size = 14)
}
\arguments{
\item{df}{Data frame with information about position of junction etc.}

\item{flanking_size}{number amino acids left and right of the breakpoint or novel sequence part}
}
\description{
Get peptide context sequence given a tibble with
\code{normalized_protein_junc_pos}
\code{is_first_reading_frame}
\code{frame_shift}
\code{additional_novel_AA}
\code{protein}
\code{cds_length_difference}
}
\keyword{internal}
