% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_junc_pos.R
\name{get_unknown_exon_intron_rentention}
\alias{get_unknown_exon_intron_rentention}
\title{Get the exon range flanking the other side of an intron retention of interest}
\usage{
get_unknown_exon_intron_rentention(tx_lst, tx_mod, jx_lst, start_on_exon)
}
\arguments{
\item{tx_lst}{transcripts a \code{\link[GenomicRanges]{GRangesList}} with
transcripts defined as GRanges of exons.}

\item{tx_mod}{modified transcripts a \code{\link[GenomicRanges]{GRangesList}} with
transcripts defined as GRanges of exons.}

\item{jx_lst}{splice junctions as GRanges objects in list format}

\item{start_on_exon}{a logical vector indicating if start or and end of junc is on exon.}
}
\value{
a GRange object with the exon range of the unknown exon
}
\description{
Get the exon range flanking the other side of an intron retention of interest
}
