% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_junc.R
\name{grl_update_start_at}
\alias{grl_update_start_at}
\title{update start coordinate of a subset of ranges by given indices}
\usage{
grl_update_start_at(grl, at, pos)
}
\arguments{
\item{grl}{A \code{\link[GenomicRanges]{GRangesList}} object}

\item{at}{The indices of ranges to update}

\item{pos}{The new coordinate}
}
\value{
A modified list object
}
\description{
update start coordinate of a subset of ranges by given indices
}
