% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irfinder_parsing.R
\name{import_irfinder_txt}
\alias{import_irfinder_txt}
\title{Imports tabular IRFinder retained introns predictions}
\usage{
import_irfinder_txt(file.name)
}
\arguments{
\item{file.name}{Path to "IR-(non)dir-(val).txt"}
}
\value{
A tibble with one retained intron call per row
}
\description{
Imports tabular IRFinder retained introns predictions
}
