% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_junctions_RNA_tools.R
\name{import_leafcutter.bam}
\alias{import_leafcutter.bam}
\title{Imports "_Aligned.out.bam.junc" file from Leafcutter output}
\usage{
import_leafcutter.bam(file.name)
}
\arguments{
\item{file.name}{The path to the file}
}
\value{
A tibble with columns "chromosome", "junction_start", "junction_end",
"dot", "count", "strand"
}
\description{
Imports "_Aligned.out.bam.junc" file from Leafcutter output
}
