% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_junctions_RNA_tools.R
\name{import_leafcutter.counts}
\alias{import_leafcutter.counts}
\title{Imports "_perind.counts.gz" from Leafcutter outpu.}
\usage{
import_leafcutter.counts(file.name)
}
\arguments{
\item{file.name}{The path to the file}
}
\value{
A tibble with columns "intron_cluster" and "counts
}
\description{
Imports "_perind.counts.gz" from Leafcutter outpu.
}
