% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafcutter_parsing.R
\name{import_leafcutter_bam}
\alias{import_leafcutter_bam}
\title{Imports "_Aligned.out.bam.junc" file from LeafCutter output}
\usage{
import_leafcutter_bam(file.name)
}
\arguments{
\item{file.name}{The path to the file}
}
\value{
A tibble with columns "chromosome", "junction_start", "junction_end",
"dot", "count", "strand"
}
\description{
Imports "_Aligned.out.bam.junc" file from LeafCutter output
}
