% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_regtools.R
\name{import_regtools_junc}
\alias{import_regtools_junc}
\title{Imports Regtools junctions annotate table}
\usage{
import_regtools_junc(file.name)
}
\arguments{
\item{file.name}{The path to the file}
}
\value{
A tibble with regtools identified junctions
}
\description{
Imports Regtools junctions annotate table
}
