% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spladder_parsing.R
\name{import_spladder}
\alias{import_spladder}
\title{Imports SplAdder output from a given path with ".confirmed.txt.gz" files.
The results for one patient should be stored in the given path. Please note
that multiple (coordinated) exons skips (mult_exon_skip) are currently not supported.}
\usage{
import_spladder(path)
}
\arguments{
\item{path}{The path to a folder with SplAdder output. This folder must contain the ".confirmed.txt.gz"
files for the alternative splicing type of interest.}
}
\value{
A list with tibbles. Each tibble is a SplAdder output for "A5SS",
"A3SS", "cassette_exon", "intron_retention", "mutex_exons".
}
\description{
Imports SplAdder output from a given path with ".confirmed.txt.gz" files.
The results for one patient should be stored in the given path. Please note
that multiple (coordinated) exons skips (mult_exon_skip) are currently not supported.
}
