% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_star.R
\name{import_star_sj}
\alias{import_star_sj}
\title{Imports tabular STAR SJ.out.tab splice junctions}
\usage{
import_star_sj(file.name)
}
\arguments{
\item{file.name}{Path to STAR SJ.out.tab}
}
\value{
A tibble with one splice junction per row and columns
chromosome, intron_start, intron_end, strand_enc, motif_enc,
annotation, uniquely_mapping_reads, multi_mapping_reads, alignment_overhang
}
\description{
Imports tabular STAR SJ.out.tab splice junctions
}
\examples{

path <-  system.file("extdata", "test_star_SJ.out.tab", package = "splice2neo")
import_star_sj(path)

}
