% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_stringtie.R
\name{import_stringtie_gtf}
\alias{import_stringtie_gtf}
\title{Imports StringTie GTF file}
\usage{
import_stringtie_gtf(file.name)
}
\arguments{
\item{file.name}{Path to StringTie GTF}
}
\value{
A tibble with one splice junction per row and columns
chromosome, junction_start, junction_end, strand, junc_id
}
\description{
Imports StringTie GTF file
}
