% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_peptide.R
\name{is_first_reading_frame}
\alias{is_first_reading_frame}
\title{Determine if transcript is in first reading frame, given a tibble with:
\code{junc_pos_cds}}
\usage{
is_first_reading_frame(df)
}
\arguments{
\item{df}{Sequence

Returns \code{TRUE} if transcript is in first reading frame and \code{FALSE} if transcript follows second or third reading frame.
occurs.}
}
\description{
Determine if transcript is in first reading frame, given a tibble with:
\code{junc_pos_cds}
}
