% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_in_rnaseq.R
\name{is_in_rnaseq}
\alias{is_in_rnaseq}
\title{test if junction was found in corresponding RNA-seq data}
\usage{
is_in_rnaseq(junc_id, rna_juncs)
}
\arguments{
\item{junc_id}{vector of junction id to test}

\item{rna_juncs}{a vector of junctions that were identified in RNA-seq data}
}
\value{
logical vector of same length as \code{junc_id} indicating
if the input \code{junc_id} was found in RNA-seq data of the same sample
}
\description{
test if junction was found in corresponding RNA-seq data
}
