% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_junc_pos.R
\name{is_start_on_exon}
\alias{is_start_on_exon}
\title{Tests if start or and end of junc is on exon.}
\usage{
is_start_on_exon(tx_lst, jx_lst, intron_retention)
}
\arguments{
\item{tx_lst}{transcripts a \code{\link[GenomicRanges]{GRangesList}} with
transcripts defined as GRanges of exons.}

\item{jx_lst}{splice junctions as GRanges objects in list format}

\item{intron_retention}{logical vector indicating if junctions belong to an intron retention event}
}
\value{
a logical vector indicating if start or and end of junc is on exon.
}
\description{
Tests if start or and end of junc is on exon.
}
