% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_junc_id.R
\name{junc2breakpoint}
\alias{junc2breakpoint}
\title{Transforms the junction id into the breakpoint id of the format \verb{<chr>:<start>-<chr>:<end>}}
\usage{
junc2breakpoint(junc_id)
}
\arguments{
\item{junc_id}{The junction id in the format \verb{<chr>:<start>-<end>:<strand>}}
}
\value{
The breakpoint kit id in the format \verb{<chr>:<start>-<chr>:<end>}
}
\description{
Transforms the junction id into the breakpoint id of the format \verb{<chr>:<start>-<chr>:<end>}
}
\examples{
junc_id <- generate_junction_id(chr = "chr1", start = 50, end = 100, strand = "+")
junc2breakpoint(junc_id)

}
