% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_canonical.R
\name{junc_id_to_gr}
\alias{junc_id_to_gr}
\title{Convert junc_id to GRanges}
\usage{
junc_id_to_gr(junc_id, use_strand = TRUE)
}
\arguments{
\item{junc_id}{junction id}

\item{use_strand}{If TRUE, strand information is used. Otherwise strand is set to "*".}
}
\value{
GRanges object of with the ranges defined in junc_id
}
\description{
Convert junc_id to GRanges
}
