% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junc_to_cts.R
\name{junc_to_cts}
\alias{junc_to_cts}
\title{Annotate splice junctions with resulting transcripts and peptide sequence}
\usage{
junc_to_cts(junc_id, transcripts, tx_id = NA, size = 400, bsg = NULL)
}
\arguments{
\item{junc_id}{junction id consisting of genomic coordinates}

\item{transcripts}{\code{\link[GenomicRanges]{GRangesList}} of transcripts}

\item{tx_id}{a vector of transcript ids in the same length as \code{junc_id}. If
provided, only transcripts with the associated id will be considered for
each junction. If NA (default) no input transcripts will be considered.}

\item{size}{the total size of the output sequence (might be shorter if transcripts is shorter)}

\item{bsg}{\code{\link[BSgenome]{BSgenome}} object such as
\code{\link[BSgenome.Hsapiens.UCSC.hg19]{BSgenome.Hsapiens.UCSC.hg19}}}
}
\value{
A data.frame with annotated input junctions with colums:
\itemize{
\item \code{junc_id} the input \code{junc_id}
\item \code{tx_id} the id of the used transcript
\item \code{junc_pos_tx} the junction position in the modified transcript sequence
\item \code{tx_id_alt} an identifier made from \code{tx_id} and \code{junc_id}
\item \code{cts_seq} the context sequence
\item \code{cts_junc_pos} the junction position in the context sequence
\item \code{cts_junc_pos} the size of the context sequence
\item \code{cts_id} a unique id for the context sequence as hash value using the
XXH128 hash algorithm
}
}
\description{
Annotate splice junctions with resulting transcripts and peptide sequence
}
\examples{

requireNamespace("BSgenome.Hsapiens.UCSC.hg19", quietly = TRUE)
bsg <- BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19

junc_to_cts(toy_junc_id, toy_transcripts, size = 400, bsg = bsg)

}
