% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junc_to_peptide.R
\name{junc_to_peptide}
\alias{junc_to_peptide}
\title{Annotate splice junctions with resulting CDS and peptide sequence}
\usage{
junc_to_peptide(junc_id, cds, tx_id = NA, size = 30, bsg = NULL)
}
\arguments{
\item{junc_id}{junction id consisting of genomic coordinates}

\item{cds}{\code{\link[GenomicRanges]{GRangesList}} of CDS}

\item{tx_id}{a vector of transcript ids in the same length as \code{junc_id}. If
provided, only transcripts with the associated id will be considered for
each junction. If NA (default) no input CDS will be considered.}

\item{size}{the total size of the output sequence (might be shorter if peptide is shorter)}

\item{bsg}{\code{\link[BSgenome]{BSgenome}} object such as
\code{\link[BSgenome.Hsapiens.UCSC.hg19]{BSgenome.Hsapiens.UCSC.hg19}}}
}
\value{
A data.frame with annotated input junctions
\itemize{
\item \code{junc_id} the input \code{junc_id}
\item \code{tx_id_input} the input \code{tx_id}
\item \code{tx_id} the id of the used cds
\item \code{tx_id_alt} a combination of the used cds id with the junction id
\item \code{peptide} the full pepetide sequence of the translated cds.
\item \code{peptide_junc_pos} The position of the junctino in the \code{peptide} sequence
\item \code{junc_in_orf} Indicator whehter the junction is located in an open reading frame.
\item \code{pep_context_seq_full} the peptide sequence around the junction includig stop codons.
\item \code{peptide_context} the peptide sequence around the junction without stop codons.
\item \code{peptide_context_junc_pos} The junction position relative to the \code{peptide_context} sequence
}
}
\description{
Annotate splice junctions with resulting CDS and peptide sequence
}
\examples{

requireNamespace("BSgenome.Hsapiens.UCSC.hg19", quietly = TRUE)
bsg <- BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19

junc_to_peptide("chr2_152389996_152392205_-", toy_cds, size = 30, bsg = bsg)

junc_to_peptide(toy_junc_id, toy_cds, tx_id = toy_junc_id_enst, size = 30, bsg = bsg)

}
