% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_context_sequence.R
\name{juncid2context}
\alias{juncid2context}
\title{Returns all context sequences that can derive from a given junction id}
\usage{
juncid2context(junc_id, transcript_db, genome_db, window_size = 200)
}
\arguments{
\item{junc_id}{A junction id}

\item{transcript_db}{a GRangesList with transcripts composed of exon ranges}

\item{genome_db}{The genome as DNAStringSet of the genomic sequences of each chromosome or a BSGenome object e.g. BSgenome.Hsapiens.UCSC.hg19}

\item{window_size}{number of nucleotides left and right from the "breakpoint"}
}
\value{
A tibble with sorted columns as given above
}
\description{
Returns all context sequences that can derive from a given junction id
}
\examples{


}
