% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_junctions_RNA_tools.R
\name{leafcutter.generate.output}
\alias{leafcutter.generate.output}
\title{Transforms Leafcutter intermediate files into standardized format}
\usage{
leafcutter.generate.output(tib)
}
\arguments{
\item{tib}{a tibble from leafcutter intermediate format}
}
\value{
A tibble in standardized junction format
}
\description{
Transforms Leafcutter intermediate files into standardized format
}
