% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{leafcutter_juncs}
\alias{leafcutter_juncs}
\title{A tibble with junction in standardized format. example data}
\format{
A tibble frame with 11 variables
}
\source{
random
}
\usage{
leafcutter_juncs
}
\description{
A tibble with junction in standardized format. example data
}
\keyword{datasets}
