% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_junctions_RNA_tools.R
\name{leafcutter_transform}
\alias{leafcutter_transform}
\title{Imports "_perind.counts.gz" from Leafcutter output and tranforms it
into standardized output format}
\usage{
leafcutter_transform(path)
}
\arguments{
\item{path}{The path to leafcutter output}
}
\value{
A tibble in standardized junction format
}
\description{
Imports "_perind.counts.gz" from Leafcutter output and tranforms it
into standardized output format
}
