% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_junctions_RNA_tools.R
\name{leafcutter_transform_format}
\alias{leafcutter_transform_format}
\title{Transforms Leafcutter intermediate files into standardized format}
\usage{
leafcutter_transform_format(tib)
}
\arguments{
\item{tib}{a tibble from leafcutter intermediate format}
}
\value{
A tibble in standardized junction format
}
\description{
Transforms Leafcutter intermediate files into standardized format
}
