% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftover_junc_id.R
\name{liftover_junc_id}
\alias{liftover_junc_id}
\title{LiftOver junction IDs using the liftOver tool}
\usage{
liftover_junc_id(junc_df, chain_file)
}
\arguments{
\item{junc_df}{a data.frame with at least one column \code{junc_id} containing junction IDs}

\item{chain_file}{path to a chain file for the UCSC liftOver tool. See also \code{\link[rtracklayer]{liftOver}}}
}
\value{
a data.frame like the input \code{junc_df} with the following additional columns:
\itemize{
\item \code{liftover_successful} a logical vector indicating if the liftOver was successful and lifted junction positions build valid genomic intervals and not single positions.
\item \code{liftover_unique} a logical vector indicating if the liftOver was unique (1-to-1 correspondence).
\item \code{junc_id_lifted_collapsed} a character vector with the lifted junction IDs.
Multiple IDs are separated by \code{|}.
NA represent junc_ids that could not be lifted.
\item \code{junc_id_lifted} a character vector with a unique lifted junction IDs.
Potentially multiple lifted IDs are combined by the minimal start and maximal
end coordinate. NA represent junc_ids that could not be lifted.
}
}
\description{
This can be used to liftOver junctions to personalized genome coordinates.
}
\examples{

chain_file = system.file(package="liftOver", "extdata", "hg38ToHg19.over.chain")
junc_df <- toy_junc_df
liftover_junc_id(junc_df, chain_file)

}
