% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_context_sequence.R
\name{map_junc_transcript}
\alias{map_junc_transcript}
\title{Returns transcripts in which alternative splicing events are predicted}
\usage{
map_junc_transcript(junc_id, junc_pos1, junc_pos2, junc_df, transcript_db)
}
\arguments{
\item{junc_id}{A junction id}

\item{junc_pos1}{GRanges object for junction position 1}

\item{junc_pos2}{GRanges object for junction position 2}

\item{junc_df}{A tibble with columns \code{junc_id}, \code{chr}, \code{pos1}, \code{pos2}, \code{strand}}

\item{transcript_db}{a GRangesList with transcripts composed of exon ranges}
}
\value{
A tibble with columns \code{junc_id}, \code{chr}, \code{pos1}, \code{pos2}, \code{strand},
\code{jidx}, \code{subjectHits}, \code{enst}. This tibble returns all transcripts that are
affected by a given junction.
}
\description{
Returns transcripts in which alternative splicing events are predicted
}
\examples{

}
