% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyquant.R
\name{map_requant}
\alias{map_requant}
\title{Maps the re-quantification result from EasyQuant. on the junction-transcript centric tibble by hash id.}
\usage{
map_requant(path_to_easyquant_folder, junc_tib)
}
\arguments{
\item{path_to_easyquant_folder}{The path to EasyQuant folder}

\item{junc_tib}{The junction-transcript centric tibble, i.e. each row represents an altered transcript. Must contain a column \code{hash_id} with hash ids that relate to the column \code{name} in the Easyquant output.}
}
\value{
Extended junction-transcript tibble with re-quantification results.  The following columns are added:
\itemize{
\item \code{interval_position}: comma-separated start end position of the interval relative to input sequence.
\item \code{overlap_interval_end_reads}: reads overlapping the the interval end.
\item \code{span_interval_end_reads}: read pairs spanning the interval end.
\item \code{within_interval_reads}: number of reads that map into the interval.
\item \code{interval_coverage_perc}: interval coverage defined as the percentage of the given interval that is covered by reads.
\item \code{interval_coverage_mean}: interval coverage defined as the mean number of reads covering a position in the given interval.
}
}
\description{
Maps the re-quantification result from EasyQuant. on the junction-transcript centric tibble by hash id.
}
\examples{


}
