% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_tx_fast.R
\name{modify_tx_fast}
\alias{modify_tx_fast}
\title{Modify transcript by introducing splice junctions}
\usage{
modify_tx_fast(tx, jx)
}
\arguments{
\item{tx}{transcripts a named \code{\link[GenomicRanges]{GRangesList}} with
transcripts defined as GRanges of exons.}

\item{jx}{splice junctions as GRanges objects with the same length as \code{tx}.}
}
\value{
a \code{\link[GenomicRanges]{GRangesList}} with altered transcripts.
}
\description{
Modify transcript by introducing splice junctions
}
\examples{
tx <- GenomicRanges::GRangesList(
  list(GenomicRanges::GRanges(c(
     "1:2-3:+",
     "1:5-6:+",
     "1:10-15:+"
  )))
)

jx <- GenomicRanges::GRanges(c("1:3-10:+"))

modify_tx(tx, jx)

}
