% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_mut_effect.R
\name{next_junctions}
\alias{next_junctions}
\title{Get genomic coordinates of possible next donor and acceptor sides.}
\usage{
next_junctions(var_gr, transcripts, transcripts_gr)
}
\arguments{
\item{var_gr}{a GenomicRanges object with variants (of length one).
It is assumed that it is named with a unique ID.
Usually this is \verb{<chr>_<pos>_<alt>}.}

\item{transcripts}{a GRangesList with transcripts defined as GRanges of exons
created by \code{GenomicFeatures::exonsBy(txdb, by = c("tx"), use.names = TRUE)}.}

\item{transcripts_gr}{a GRanges object with transcript ranges created by
\code{GenomicFeatures::transcripts(txdb)}}
}
\value{
A data.frame with possible upstream and downstream exon coordinates for
all overlapping transcripts. The data frame contains the following columns:
mut_id, var_nr, tx_chr, tx_id, exon_idx, tx_strand, upstream_start,
upstream_end, downstream_start, downstream_end, at_start, at_end.
}
\description{
Get genomic coordinates of possible next donor and acceptor sides.
}
