% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_pangolin.R
\docType{data}
\name{pangolin_effect_translation}
\alias{pangolin_effect_translation}
\title{This dataset translates the increase/decrease splicing score from pangolin
into a donor gain/loss and acceptor gain/loss effect annotations.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 2 columns.
}
\usage{
pangolin_effect_translation
}
\description{
This dataset translates the increase/decrease splicing score from pangolin
into a donor gain/loss and acceptor gain/loss effect annotations.
}
\keyword{internal}
