% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_cispliceai.R
\name{parse_cispliceai_thresh}
\alias{parse_cispliceai_thresh}
\title{Parse VCF output file from CI-SpliceAI with -t flag as table}
\usage{
parse_cispliceai_thresh(vcf_file)
}
\arguments{
\item{vcf_file}{path to a single VCF file, the output of spliceAI}
}
\value{
a \link[tibble:tibble-package]{tibble} with one row per variant
annotation. Each input variant can have multiple annotations.
}
\description{
requires the vcfR package.
}
\examples{

cispliceai_file <- system.file("extdata", "cispliceai_thresh_output.vcf", package = "splice2neo")
parse_cispliceai(cispliceai_file)

}
\seealso{
\code{\link{format_cispliceai}}, \code{\link{annotate_mut_effect}}
}
