% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irfinder_parsing.R
\name{parse_irfinder_txt}
\alias{parse_irfinder_txt}
\title{Imports "IRFinder-IR-nondir.txt" from IRFinder short mode
and transforms the raw output into standardized junction output format.}
\usage{
parse_irfinder_txt(path, warnings = FALSE, irratio = 0.1, cnn = FALSE)
}
\arguments{
\item{path}{The path to "IRFinder-IR-nondir.txt"}

\item{warnings}{Filter out IR calls with Warnings}

\item{irratio}{Filter out IR calls with IRRatio <= this parameter}

\item{cnn}{Use CNN validated table as input instead of standard output table}
}
\value{
A tibble in standardized junction format
}
\description{
Imports "IRFinder-IR-nondir.txt" from IRFinder short mode
and transforms the raw output into standardized junction output format.
}
