% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmsplice.R
\name{parse_mmsplice}
\alias{parse_mmsplice}
\title{Parse .csv file from MMsplice output as data.frame}
\usage{
parse_mmsplice(infile)
}
\arguments{
\item{infile}{path to a .csv file, the output of MMsplice}
}
\value{
a \link[tibble:tibble-package]{tibble} with one row per variant and
affected exon.
}
\description{
Parse .csv file from MMsplice output as data.frame
}
\examples{

in_file <- system.file("extdata", "mmsplice_pred.csv", package = "splice2neo")
parse_mmsplice(in_file)

}
