% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_pangolin.R
\name{parse_pangolin}
\alias{parse_pangolin}
\title{Parse VCF output file from pangolin as table}
\usage{
parse_pangolin(vcf_file)
}
\arguments{
\item{vcf_file}{path to a single VCF file, the output of pangolin}
}
\value{
a \link[tibble:tibble-package]{tibble} with one row per variant
annotation. Each input variant can have multiple annotations.
}
\description{
requires the vcfR package.
}
\examples{

pangolin_file <- system.file("extdata", "spliceai_output.pangolin.vcf", package = "splice2neo")
parse_pangolin(pangolin_file)

}
