% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_spliceai.R
\name{parse_spliceai}
\alias{parse_spliceai}
\title{Parse VCF output file from spliceAI as table}
\usage{
parse_spliceai(vcf_file)
}
\arguments{
\item{vcf_file}{path to a single VCF file, the output of spliceAI}
}
\value{
a \link[tibble:tibble-package]{tibble} with one row per variant
annotation. Each input variant can have multiple annotations.
}
\description{
requires the vcfR package.
}
\examples{

spliceai_file <- system.file("extdata", "spliceai_output.vcf", package = "splice2neo")
parse_spliceai(spliceai_file)

}
