% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_spliceai.R
\name{parse_spliceai_thresh}
\alias{parse_spliceai_thresh}
\title{Parse VCF output file from SpliceAI with -T flag as table}
\usage{
parse_spliceai_thresh(vcf_file)
}
\arguments{
\item{vcf_file}{path to a single VCF file, the output of spliceAI}
}
\value{
a \link[tibble:tibble-package]{tibble} with one row per variant
annotation. Each input variant can have multiple annotations.
}
\description{
requires the vcfR package.
}
\examples{

spliceai_file <- system.file("extdata", "spliceai_thresh_output.vcf", package = "splice2neo")
parse_spliceai_thresh(spliceai_file)

}
\seealso{
\code{\link{format_spliceai_thresh}}, \code{\link{annotate_mut_effect}}
}
