% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyquant.R
\name{read_requant}
\alias{read_requant}
\title{Imports the re-quantification results from analysis with EasyQuant}
\usage{
read_requant(path_folder)
}
\arguments{
\item{path_folder}{The path to the EasyQuant folder}
}
\value{
A tibble with with the re-quantification results. This tibble has the columns
\itemize{
\item \code{name}: name of the input sequence.
\item \code{interval_position}: comma-separated start end position of the interval relative to input sequence.
\item \code{overlap_interval_end_reads}: reads overlapping the the interval end.
\item \code{span_interval_end_reads}: read pairs spanning the interval end.
\item \code{within_interval_reads}: number of reads that map into the interval.
\item \code{interval_coverage_perc}: interval coverage defined as the percentage of the given interval that is covered by reads.
\item \code{interval_coverage_mean}: interval coverage defined as the mean number of reads covering a position in the given interval.
}
}
\description{
Imports the re-quantification results from analysis with EasyQuant
}
\examples{


}
