% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_suppa2.R
\name{read_suppa_ioe}
\alias{read_suppa_ioe}
\title{Imports "_strict.ioe" from SUPPA2 output.}
\usage{
read_suppa_ioe(file.name)
}
\arguments{
\item{file.name}{The path to the file}
}
\value{
A tibble with columns "seqnames", "gene_id" "event_id"
"inclusion_transcripts" and "total_transcripts"
}
\description{
Imports "_strict.ioe" from SUPPA2 output.
}
