% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_regtools.R
\name{regtools_transform}
\alias{regtools_transform}
\title{Imports Regtools junctions annotate output and transforms the raw output
into standardized junction output format}
\usage{
regtools_transform(path)
}
\arguments{
\item{path}{The path to the Regtools output file}
}
\value{
A tibble in standardized junction format
}
\description{
\itemize{
\item GitHub: https://github.com/griffithlab/regtools
\item Paper: https://doi.org/10.1038/s41467-023-37266-6
}
}
\details{
Ignores junctions without strand annotation.
}
