% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junc_to_peptide.R
\name{seq_extract_nonstop}
\alias{seq_extract_nonstop}
\title{Extract sequence at position from after previous stop codon and before next.}
\usage{
seq_extract_nonstop(seq, pos)
}
\description{
@param seq Sequence
@param pos position relative to sequence
}
\details{
@return a data.frame

@examples
seq <-  "QIP\emph{LGSNSLLFPYQLMAGSTRP}SWALGC"
seq <-  c(
"QIP\emph{LGSNSLLFPYQLMAGSTRP}SWALGC",
"LKMRGDTNDILSHLD\emph{REQRVGQ}AEAASP"
)
pos <- c(14, 14)
seq_extract_nonstop(seq, pos)
}
