% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junc_to_peptide.R
\name{seq_truncate_nonstop}
\alias{seq_truncate_nonstop}
\title{Truncate input sequence after input position before next stop codons (\code{*}).}
\usage{
seq_truncate_nonstop(seq, pos)
}
\arguments{
\item{seq}{Sequence}

\item{pos}{position relative to sequence}
}
\value{
a character
}
\description{
Truncate input sequence after input position before next stop codons (\code{*}).
}
\examples{

seq_truncate_nonstop("1234*6789", 2) # "1234"
seq_truncate_nonstop("1234*6789", 8) # "1234*6789"

seq <- "QIP*LGSNSLLFPYQLMAGSTRP*SWALGC"
seq_truncate_nonstop(seq, 14) #"QIP*LGSNSLLFPYQLMAGSTRP"

}
