% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spladder_parsing.R
\name{sort_columns}
\alias{sort_columns}
\title{Sorts columns of junction output file in the following order:
"junction_start", "junction_end", "strand", "chromosome", "Gene",
"class", "AS_event_ID", "junction_id"}
\usage{
sort_columns(tib)
}
\arguments{
\item{tib}{A tibble with the following as given in the description.}
}
\value{
A tibble with sorted columns as given above
}
\description{
Sorts columns of junction output file in the following order:
"junction_start", "junction_end", "strand", "chromosome", "Gene",
"class", "AS_event_ID", "junction_id"
}
