% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_junctions_RNA_tools.R
\name{spladder.transform.a5ss}
\alias{spladder.transform.a5ss}
\title{Transforms events from alternative 5' splice sites from SPLADDER output
format into standardized junction format}
\usage{
spladder.transform.a5ss(tib)
}
\arguments{
\item{tib}{A tibble in SPLADDER output format}
}
\value{
A tibble in standardized junction format
}
\description{
Transforms events from alternative 5' splice sites from SPLADDER output
format into standardized junction format
}
\examples{
spladder_output.a5ss
transformed_spladder.a5ss <- spladder.transform.a5ss(spladder_output.a5ss)
transformed_spladder.a5ss

}
