% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_junctions_RNA_tools.R
\name{spladder.transform.format}
\alias{spladder.transform.format}
\title{Transforms SPLADDER output into standardized junction format}
\usage{
spladder.transform.format(l)
}
\arguments{
\item{l}{A list with tibbles. Each tibble is a Spladder output for "A5SS",
"A3SS", "cassette_exon", "intron_retention", "mutex_exons".}
}
\value{
A tibble in standardized junction format, combining all alternative
splicing classes that are covered by Spladder
}
\description{
Transforms SPLADDER output into standardized junction format
}
\examples{
spladder_output
transformed_spladder <- spladder.transform.format(
 spladder_output)
transformed_spladder

}
