% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spladder_output.a5ss}
\alias{spladder_output.a5ss}
\title{A tibble containing events from alternative 5' splice sites identified by
SplAdder}
\format{
A tibble frame with 1 and 15 variables:
}
\source{
random
}
\usage{
spladder_output.a5ss
}
\description{
A tibble containing events from alternative 5' splice sites identified by
SplAdder
}
\keyword{datasets}
