% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spladder_output.exonskip}
\alias{spladder_output.exonskip}
\title{A tibble containing events from exon skipping identified by
SplAdder}
\format{
A tibble frame with 1 and 15 variables:
}
\source{
random
}
\usage{
spladder_output.exonskip
}
\description{
A tibble containing events from exon skipping identified by
SplAdder
}
\keyword{datasets}
