% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spladder_output.intronreten}
\alias{spladder_output.intronreten}
\title{A tibble containing events from intron retention identified by
Spladder}
\format{
A tibble frame with 1 and 15 variables:
}
\source{
random
}
\usage{
spladder_output.intronreten
}
\description{
A tibble containing events from intron retention identified by
Spladder
}
\keyword{datasets}
