% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spladder_output.mutexon}
\alias{spladder_output.mutexon}
\title{A tibble containing events from mutually exclusive exons identified by
SplAdder}
\format{
A tibble frame with 1 row and 17 variables:
}
\source{
random
}
\usage{
spladder_output.mutexon
}
\description{
A tibble containing events from mutually exclusive exons identified by
SplAdder
}
\keyword{datasets}
