% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spladder_output}
\alias{spladder_output}
\title{A list of tibbles with each list containing events from a different
splicing class identified by Spladder}
\format{
A list frame with tibbles called: A3SS, A5SS, casette_exon,
intron_retention and mutex_exons
}
\source{
random
}
\usage{
spladder_output
}
\description{
A list of tibbles with each list containing events from a different
splicing class identified by Spladder
}
\keyword{datasets}
