% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_junctions_RNA_tools.R
\name{spladder_transform}
\alias{spladder_transform}
\title{Imports SplAdder output from a given path and transforms it into standardized
junction format}
\usage{
spladder_transform(path)
}
\arguments{
\item{path}{The path to a folder with SplAdder output}
}
\value{
A tibble in standardized junction format, combining all alternative
splicing classes that are covered by SplAdder
}
\description{
Imports SplAdder output from a given path and transforms it into standardized
junction format
}
