% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_junctions_RNA_tools.R
\name{spladder_transform_a3ss}
\alias{spladder_transform_a3ss}
\title{Transforms events from alternative 3' splice sites from SPLADDER output
format into standardized junction format}
\usage{
spladder_transform_a3ss(tib)
}
\arguments{
\item{tib}{A tibble in SPLADDER output format}
}
\value{
A tibble in standardized junction format
}
\description{
Transforms events from alternative 3' splice sites from SPLADDER output
format into standardized junction format
}
\examples{
spladder_output.a3ss
transformed_spladder.a3ss <- spladder_transform_a3ss(spladder_output.a3ss)
transformed_spladder.a3ss

}
