% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_junctions_RNA_tools.R
\name{spladder_transform_a5ss}
\alias{spladder_transform_a5ss}
\title{Transforms events from alternative 5' splice sites from SplAdder output
format into standardized junction format}
\usage{
spladder_transform_a5ss(tib)
}
\arguments{
\item{tib}{A tibble in SplAdder output format}
}
\value{
A tibble in standardized junction format
}
\description{
Transforms events from alternative 5' splice sites from SplAdder output
format into standardized junction format
}
\examples{
spladder_output.a5ss
transformed_spladder.a5ss <- spladder_transform_a5ss(spladder_output.a5ss)
transformed_spladder.a5ss

}
