% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spladder_parsing.R
\name{spladder_transform_ass}
\alias{spladder_transform_ass}
\title{Transforms events from alternative 3' or 5' splice sites from SplAdder output
format into standardized junction format}
\usage{
spladder_transform_ass(tib, AS_class)
}
\arguments{
\item{tib}{A tibble in SplAdder output format}

\item{AS_class}{The type of alternative splice site. Can be "A3SS" or "A5SS"}
}
\value{
A tibble in standardized junction format
}
\description{
Transforms events from alternative 3' or 5' splice sites from SplAdder output
format into standardized junction format
}
