% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spladder_parsing.R
\name{spladder_transform_format}
\alias{spladder_transform_format}
\title{Transforms SplAdder output into standardized junction format}
\usage{
spladder_transform_format(l)
}
\arguments{
\item{l}{A list with tibbles that contain the SplAdder output - each for one type of alternative splicing. These types can be "A5SS",
"A3SS", "cassette_exon", "intron_retention", "mutex_exons".}
}
\value{
A tibble in standardized junction format, combining all alternative
splicing classes that are were determined with SplAdder
}
\description{
Transforms SplAdder output into standardized junction format
}
\examples{
spladder_output
transformed_spladder <- spladder_transform_format(
 spladder_output)
transformed_spladder

}
