% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_junctions_RNA_tools.R
\name{spladder_transform_mutex_exon}
\alias{spladder_transform_mutex_exon}
\title{Transforms events resulting from mutually exclusive exons from SPLADDER output
format into standardized junction format}
\usage{
spladder_transform_mutex_exon(tib)
}
\arguments{
\item{tib}{A tibble in SPLADDER output format}
}
\value{
A tibble in standardized junction format
}
\description{
Transforms events resulting from mutually exclusive exons from SPLADDER output
format into standardized junction format
}
\examples{
spladder_output.mutexon
transformed_spladder.mutexon <- spladder_transform_mutex_exon(
 spladder_output.mutexon)
transformed_spladder.mutexon

}
