% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_parsing.R
\name{star_transform}
\alias{star_transform}
\title{Imports "*SJ.out.tab" from STAR and transforms the raw output
into standardized junction output format.}
\usage{
star_transform(path)
}
\arguments{
\item{path}{The path to STAR SJ.out.tab}
}
\value{
A tibble in standardized junction format
}
\description{
Imports "*SJ.out.tab" from STAR and transforms the raw output
into standardized junction output format.
}
\examples{
path <-  system.file("extdata", "test_star_SJ.out.tab", package = "splice2neo")
star_juncs <- star_transform(path)
star_juncs

}
