% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_stringtie.R
\name{stringtie_transform_format}
\alias{stringtie_transform_format}
\title{Transforms StringTie intermediate table into standardized junction format}
\usage{
stringtie_transform_format(tib)
}
\arguments{
\item{tib}{StringTie junctions as tibble}
}
\value{
A tibble in standardized junction format
}
\description{
Transforms StringTie intermediate table into standardized junction format
}
