% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_suppa2.R
\name{suppa_import}
\alias{suppa_import}
\title{Imports SUPPA2 output from a given path with "_strict.ioe" files.
The results for one patient should be stored in the given path. Please note
that AF and AL events are not supported.}
\usage{
suppa_import(path)
}
\arguments{
\item{path}{The path to a folder with SUPPA2 output. This folder must contain the "_strict.ioe"
files for the alternative splicing type of interest.}
}
\value{
A list with tibbles. Each tibble is a SUPPA2 output for "A5SS",
"A3SS", "cassette_exon", "intron_retention", "mutex_exons".
}
\description{
Imports SUPPA2 output from a given path with "_strict.ioe" files.
The results for one patient should be stored in the given path. Please note
that AF and AL events are not supported.
}
