% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_junc_df}
\alias{toy_junc_df}
\title{An example dataset of 18 splice junctions from human
(hg19) genome as a data.frame}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 17 rows and 2 columns.
}
\source{
They were created from the spliceAI example file
}
\usage{
toy_junc_df
}
\description{
The data.frame has the following columns:
}
\details{
\itemize{
\item \code{junc_id} see \link{toy_junc_id}
\item \code{tx_id} see \link{toy_junc_id_enst}
}
}
\keyword{datasets}
