% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_junc_id}
\alias{toy_junc_id}
\title{An example dataset of 18 splice junctions in \code{junc_id} format from human
(hg19) genome as character vector}
\format{
A character vector with \code{junc_id}s in the format:
\verb{<chr>_<pos1>_<pos2>_<strand>}
}
\source{
They were created from the spliceAI example file
}
\usage{
toy_junc_id
}
\description{
An example dataset of 18 splice junctions in \code{junc_id} format from human
(hg19) genome as character vector
}
\keyword{datasets}
