% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_junc_id_enst}
\alias{toy_junc_id_enst}
\title{An example dataset of 18 transcript IDs matching to the junctions in \code{toy_junc_id}}
\format{
A character vector with ENSEMBL transcript ids
}
\source{
They were created using transcript annotations from
\url{https://www.ensembl.org}
}
\usage{
toy_junc_id_enst
}
\description{
An example dataset of 18 transcript IDs matching to the junctions in \code{toy_junc_id}
}
\keyword{datasets}
