% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_transcripts_gr}
\alias{toy_transcripts_gr}
\title{An example dataset of full transcript ranges from a subset of human
chromosome 2.}
\format{
A GRanges object with 1951 regions.
}
\source{
\url{ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_34/GRCh37_mapping/gencode.v34lift37.annotation.gtf.gz}
}
\usage{
toy_transcripts_gr
}
\description{
This is a GRanges object with 1951 full transcript ranges.
This data set is a subset of annotations of human (hg19) chromosome 2
restricted to this region: chr2:152000000-180000000 .
}
\keyword{datasets}
