% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyquant.R
\name{transform_for_requant}
\alias{transform_for_requant}
\title{Create table for re-quantification analysis with easyquant}
\usage{
transform_for_requant(cts_id, cts_seq, junc_position)
}
\arguments{
\item{cts_id}{An id for the cts sequence to be re-quantified. If the
context sequence was determined with the function \code{junct_to_cts} the associated hash id \code{cts_id} can be used.}

\item{cts_seq}{The context sequence.}

\item{junc_position}{The position of the junction within the context sequence.  If the
context sequence was determined with \code{junct_to_cts} the associated hash id \code{cts_junc_pos} can be used.}
}
\value{
A tibble that can be used for requantification with easyquant. This tibble has the columns
\itemize{
\item \code{name}: Name of the input sequence.
\item \code{sequence}: The context sequence that will be requantified.
\item \code{position}: The position of the junction within the context sequence.
}
}
\description{
Create table for re-quantification analysis with easyquant
}
\examples{


}
