% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irfinder_parsing.R
\name{transform_irfinder_txt}
\alias{transform_irfinder_txt}
\title{Transforms IRFinder intermediate table into standardized junction format}
\usage{
transform_irfinder_txt(tib)
}
\arguments{
\item{tib}{IRFinder IR-(non)dir-(val).txt as tibble}
}
\value{
A tibble in standardized junction format
}
\description{
Transforms IRFinder intermediate table into standardized junction format
}
