% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_junctions_RNA_tools.R
\name{transform_leafcutter.bam}
\alias{transform_leafcutter.bam}
\title{Transforms Leafcutter bam junc file}
\usage{
transform_leafcutter.bam(tib)
}
\arguments{
\item{tib}{Leafcutter bam junc file as tibble}
}
\value{
A tibble including a column with the junction id
}
\description{
Transforms Leafcutter bam junc file
}
