% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_junctions_RNA_tools.R
\name{transform_leafcutter.counts}
\alias{transform_leafcutter.counts}
\title{Transforms Leafcutter counts file}
\usage{
transform_leafcutter.counts(tib)
}
\arguments{
\item{tib}{Leafcutter counts file as tibble}
}
\value{
A tibble with columns "chromosome", "junction_start", "junction_end",
"cluster"
}
\description{
Transforms Leafcutter counts file
}
