% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_junctions_RNA_tools.R
\name{transform_leafcutter_bam}
\alias{transform_leafcutter_bam}
\title{Transforms LeafCutter bam junc file}
\usage{
transform_leafcutter_bam(tib)
}
\arguments{
\item{tib}{LeafCutter bam junc file as tibble}
}
\value{
A tibble including a column with the junction id
}
\description{
Transforms LeafCutter bam junc file
}
