% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_regtools.R
\name{transform_regtools_junc}
\alias{transform_regtools_junc}
\title{Transforms Regtools intermediate files into standardized format}
\usage{
transform_regtools_junc(tib)
}
\arguments{
\item{tib}{A tibble from Regtools intermediate format}
}
\value{
A tibble in standardized junction format
}
\description{
Transforms Regtools intermediate files into standardized format
}
