% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_parsing.R
\name{transform_star_sj}
\alias{transform_star_sj}
\title{Transforms STAR intermediate table into standardized junction format}
\usage{
transform_star_sj(tib)
}
\arguments{
\item{tib}{STAR *SJ.out.tab as tibble}
}
\value{
A tibble in standardized junction format
}
\description{
Transforms STAR intermediate table into standardized junction format
}
