% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_suppa2.R
\name{transform_suppa_ass_events}
\alias{transform_suppa_ass_events}
\title{Transforms events from alternative 3' or 5' splice sites from SUPPA2 output
format into standardized junction format}
\usage{
transform_suppa_ass_events(tib, type = "A5SS")
}
\arguments{
\item{tib}{A tibble in SUPPA2 output format}

\item{type}{The type of alternative splice site. Can be "A3SS" or "A5SS"}
}
\value{
A tibble in standardized junction format
}
\description{
Transforms events from alternative 3' or 5' splice sites from SUPPA2 output
format into standardized junction format
}
