% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_suppa2.R
\name{transform_suppa_se_events}
\alias{transform_suppa_se_events}
\title{Transforms events resulting from exon skipping from SUPPA2 output
format into standardized junction format}
\usage{
transform_suppa_se_events(tib)
}
\arguments{
\item{tib}{A tibble in SUPPA2 output format}
}
\value{
A tibble in standardized junction format
}
\description{
Transforms events resulting from exon skipping from SUPPA2 output
format into standardized junction format
}
