% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_mut_junc.R
\name{unique_junc_mmsplice}
\alias{unique_junc_mmsplice}
\title{Filter junction datasets for unique \code{mut_id},
\code{junc_id}, \code{tx_id}. If there were multiple exon inclusion events
predicted to lead to the same combination of \code{mut_id}, \code{junc_id}, \code{tx_id},
only the  effect with the maximal score is returned.}
\usage{
unique_junc_mmsplice(junc_data)
}
\arguments{
\item{junc_data}{A tibble with junctions. The tibble should contain the columns \code{mut_id},
\code{tx_id}, \code{junc_id} and \code{delta_logit_psi} and might have individual sets of other tool/source
specific columns. This tibble can be created with \code{annotate_mmsplice()}.}
}
\value{
A filtered tibble with unique junctions based on the columns
\code{mut_id}, \code{tx_id}, \code{junc_id}. The event with the maximal
\code{delta_logit_psi} is kept.
}
\description{
Filter junction datasets for unique \code{mut_id},
\code{junc_id}, \code{tx_id}. If there were multiple exon inclusion events
predicted to lead to the same combination of \code{mut_id}, \code{junc_id}, \code{tx_id},
only the  effect with the maximal score is returned.
}
