% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_mut_junc.R
\name{unique_mut_junc}
\alias{unique_mut_junc}
\title{Filter junctions for unique combinations of mutation, transcript, and
junctions.}
\usage{
unique_mut_junc(junc_data)
}
\arguments{
\item{junc_data}{A tibble with junctions. The tibble should contain the columns \code{mut_id},
\code{tx_id}, \code{junc_id} and \code{score} and might have individual sets of other tool/source
specific columns. This tibble can be created with \code{annotate_mut_effect()}.}
}
\value{
A filtered tibble with unique junctions based on the columns
\code{mut_id}, \code{tx_id}, \code{junc_id}. The event with the maximal
\code{score} is kept.
}
\description{
Filters for unqiue \code{mut_id},\code{junc_id}, \code{tx_id}. If there are multiple
effects (DL/AL) predicted to lead to the same combination of \code{mut_id},
\code{junc_id}, \code{tx_id}, only the  effect with the maximal score is returned.
}
