% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junc_to_tx.R
\name{update_end_at}
\alias{update_end_at}
\title{update end coordinate of a subset of ranges by given indices}
\usage{
update_end_at(gr, pos, at)
}
\arguments{
\item{gr}{A GRanges object}

\item{pos}{The new coordinate}

\item{at}{The indices of ranges to update}
}
\value{
A modified GRanges object
}
\description{
update end coordinate of a subset of ranges by given indices
}
