% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junc_to_tx.R
\name{update_start_at}
\alias{update_start_at}
\title{update start coordinate of a subset of ranges by given indices}
\usage{
update_start_at(gr, pos, at)
}
\arguments{
\item{gr}{A GRanges objec}

\item{pos}{The new coordinate}

\item{at}{The indeces of ranges to update}
}
\value{
A modified GRanges object
}
\description{
update start coordinate of a subset of ranges by given indices
}
