% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tx.R
\name{add_tx}
\alias{add_tx}
\title{Annotate splice junctions with affected transcript IDs}
\usage{
add_tx(df, transcripts)
}
\arguments{
\item{df}{A data.frame with splice junctions in rows and at least the colums:
\itemize{
\item \code{junc_id} junction id consisting of genomic coordinates
}}

\item{transcripts}{\code{\link[GenomicRanges]{GRangesList}} of transcripts}
}
\value{
A data.frame as the input but with potentially multiple rows
and with the additional column(s):
\itemize{
\item \code{tx_id} the ID of the affected transcript
\item \code{tx_lst} a list of \code{\link[GenomicRanges]{GRanges}} with the transcript
}
}
\description{
Annotate splice junctions with affected transcript IDs
}
\examples{

}
