% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_tx.R
\name{annotate_exon_idx}
\alias{annotate_exon_idx}
\title{Annotate splice junctions- transcripts combination and transcripts with exon details}
\usage{
annotate_exon_idx(df)
}
\arguments{
\item{df}{A data.frame with junctions- transcripts combination in rows.
\itemize{
\item \code{junc_id} junction id consisting of genomic coordinates
\item \code{tx_id} transcript id
\item \code{tx_lst} a list of \code{\link[GenomicRanges]{GRanges}} with the transcript
}

and additional columns.}
}
\value{
A data.frame as the input but with more annotations in the following columns:
\itemize{
\item \code{pos_left_on_exon_end} exon index if junction start matches with exon end coordinate
\item \code{pos_right_on_exon_start}  exon index if junction end matches with exon start coordinate
\item \code{pos_left_on_exon_idx} exon index in transcript if junction start overlaps exon
\item \code{pos_right_on_exon_idx} exon index in transcript if junction end overlaps exon
}
}
\description{
Annotate splice junctions- transcripts combination and transcripts with exon details
}
\keyword{internal}
