% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_peptide.R
\name{annotate_junc_in_orf}
\alias{annotate_junc_in_orf}
\title{Tests if junction position is in ORF
\code{normalized_protein_junc_pos}
\code{protein_len}
\code{protein}}
\usage{
annotate_junc_in_orf(df)
}
\arguments{
\item{df}{tibble}
}
\description{
Tests if junction position is in ORF
\code{normalized_protein_junc_pos}
\code{protein_len}
\code{protein}
}
\keyword{internal}
