% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_junc_id.R
\name{breakpoint2junc}
\alias{breakpoint2junc}
\title{Given the strand Transforms the breakpoint id into the junction id of the format \verb{<chr>:<start>-:<end>:<strand>}}
\usage{
breakpoint2junc(breakpoint_id, strand)
}
\arguments{
\item{breakpoint_id}{The junction id in the format \verb{<chr>:<start>-<chr>:<end>}}

\item{strand}{Strand information. "+" or "-"}
}
\value{
The junction id in the format \verb{<chr>:<start>-<end>:<strand>}
}
\description{
Given the strand Transforms the breakpoint id into the junction id of the format \verb{<chr>:<start>-:<end>:<strand>}
}
\examples{

breakpoint_id <- "chr1:500000-chr1:1000000"
breakpoint2junc(breakpoint_id = breakpoint_id, strand = "+")

}
