% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical_junctions.R
\name{canonical_junctions}
\alias{canonical_junctions}
\title{Build canonical junctions from transcripts}
\usage{
canonical_junctions(tx)
}
\arguments{
\item{tx}{a \code{\link[GenomicRanges]{GRangesList}} of reference transcripts}
}
\value{
a character vector of canonical splice junction ids

We build all canonical splice junctions that are in the annotated input transcripts.
The following lists implementation rules for adjacent canonical exon-exon junction:
\itemize{
\item strand = +: \eqn{e_i}, \eqn{s_{i+1}}
\item strand = -: \eqn{e_{i+1}}, \eqn{s_{i}}
}

We also include canonical intron-retention junctions. These are 5' donor or
3' acceptor sites of canonical exon-exon junctions that are not used in all
isoforms of the gene. They are located within an exon of other transcripts.
Canonical intron-retention junctions are defined by the coordinate of the
last exon base and the next base.
Therefore, we just need to check whether both bases are included in a single exon.
}
\description{
Build canonical junctions from transcripts
}
\examples{
gtf_file <- system.file("extdata","GTF_files","Aedes_aegypti.partial.gtf",
  package="splice2neo")

tx <- parse_gtf(gtf_file)

canonical_junctions(tx[1:10])

}
