% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_tx.R
\name{choose_tx}
\alias{choose_tx}
\title{Select a subset of transcripts per junction that are more likely to be affected by a junction.}
\usage{
choose_tx(df)
}
\arguments{
\item{df}{A data.frame with splice junctions in rows and at least the columns:
\itemize{
\item \code{junc_id} junction id consisting of genomic coordinates
\item \code{tx_id} transcript id of possibly affected transcripts
\item \code{tx_lst} a list of \code{\link[GenomicRanges]{GRanges}} with the transcript
}}
}
\value{
A data.frame as with relevant transcript and junction combinations. If tx_id is NA in the input data.frame such rows are removed from the output data.frame.

This function selects transcripts that are more likely to be affected to reduce the amount of junction and transcript combinations.
The function excludes transcripts for which both junction positions are located in an intron. Junctions in a given transcript must either represent an
exon skipping, intron retention, exitron, or alternative splice site event or have both junction positions in an exon. Other junction-transcript combinations are also excluded.
This function may loose relevant or keep irrelevant junction-transcripts in particular in regions with mutliple isoforms with distinct splicing pattern.
}
\description{
All possible affected transcripts are required and can be annotated with \code{add_tx()}
}
\examples{
junc_df <- tibble::tibble(
  junc_id = c("chr2:152389996-152392205:-", "chr2:152389996-152390729:-",
              "chr2:152389955-152389956:-")
)

junc_df <- add_tx(junc_df, toy_transcripts)
choose_tx(junc_df)

}
\seealso{
\code{\link{add_tx}}
}
