% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutated_transcript_ranges.R
\name{construct_mutated_range}
\alias{construct_mutated_range}
\title{returns genomic range of mutated transcript according to the splice event}
\usage{
construct_mutated_range(
  exon1_index,
  exon2_index,
  wt_transcript_range,
  strand_direction,
  junction_start,
  junction_end
)
}
\arguments{
\item{exon1_index}{index of exon related to first position of junction in
transcript ranges}

\item{exon2_index}{index of exon related to second position of junction in
transcript ranges}

\item{wt_transcript_range}{Genomic range of wildtype transcript in which the
alternative splicing event is taking place}

\item{strand_direction}{strand direction. Shall be "+" or "-"}

\item{junction_start}{junction start coordinate}

\item{junction_end}{junction end coordinate}
}
\value{
A tibble with columns \code{junc_id}, \code{chr}, \code{pos1}, \code{pos2}, \code{strand},
\code{jidx}, \code{subjectHits}, \code{enst}. This tibble returns all transcripts that are
affected by a given junction.
}
\description{
returns genomic range of mutated transcript according to the splice event
}
