% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_cispliceai.R
\name{format_cispliceai_thresh}
\alias{format_cispliceai_thresh}
\title{Formats CI-SpliceAI with -t flag output and filter for predicted effects}
\usage{
format_cispliceai_thresh(cispliceai_variants)
}
\arguments{
\item{spliceai_variants}{\link[tibble:tibble-package]{tibble} with parsed
spliceAI mutations from \code{\link{parse_spliceai}}}
}
\value{
A \link[tibble:tibble-package]{tibble} with splicing effects per row
}
\description{
Reformat the data for each annotated effect per row, filters effects
to have a probability score not NA and score > 0, and removes gene symbol from
data to make non-redundant output.
}
\examples{
cispliceai_file <- system.file("extdata", "cispliceai_output.vcf", package = "splice2neo")
df <- parse_cispliceai_thresh(cispliceai_file)
format_cispliceai_thresh(df)

}
\seealso{
\code{\link{parse_cispliceai_thresh}}, \code{\link{annotate_mut_effect}}
}
