% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_pangolin.R
\name{format_pangolin}
\alias{format_pangolin}
\title{Formats pangolin output and filter for predicted effects}
\usage{
format_pangolin(variants)
}
\arguments{
\item{variants}{\link[tibble:tibble-package]{tibble} with parsed
pangolin mutations from \code{\link{parse_pangolin}}}
}
\value{
A \link[tibble:tibble-package]{tibble} with splicing effects per row
}
\description{
Reformat the data for each annotated effect per row. As pangolin does not
provide information on donor or acceptor, use both for gain and loss annotations.
Effects with score = 0 are filtered out.
}
\examples{
pangolin_file <- system.file("extdata", "spliceai_output.pangolin.vcf", package = "splice2neo")
df <- parse_pangolin(pangolin_file)
format_pangolin(df)

}
\seealso{
\code{\link{parse_spliceai}}, \code{\link{annotate_mut_effect}}, \code{\link{format_spliceai}}
}
