% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_spliceai.R
\name{format_spliceai}
\alias{format_spliceai}
\title{Formats spliceAI output and filter for predicted effects}
\usage{
format_spliceai(spliceai_variants, gene_table = NULL)
}
\arguments{
\item{spliceai_variants}{\link[tibble:tibble-package]{tibble} with parsed
spliceAI mutations from \code{\link{parse_spliceai}}}

\item{gene_table}{optional \link[tibble:tibble-package]{tibble} with the columns:
\itemize{
\item \code{gene_id}: ENSEMBL gene id
\item \code{gene_name}: gene symbol
If \code{gene_table} is provided, the formatted data contains a column with the \code{gene_id}.
}}
}
\value{
A \link[tibble:tibble-package]{tibble} with splicing effects per row
}
\description{
Reformat the data for each annotated effect per row, filters effects
to have a probability score not NA and score > 0, and removes gene symbol from
data to make non-redundant output.
}
\examples{
spliceai_file <- system.file("extdata", "spliceai_output.vcf", package = "splice2neo")
df <- parse_spliceai(spliceai_file)
format_spliceai(df)

}
\seealso{
\code{\link{parse_spliceai}}, \code{\link{annotate_mut_effect}}
}
