% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_junctions_RNA_tools.R
\name{generate_combined_dataset}
\alias{generate_combined_dataset}
\title{Combines tibbles with junctions from Spladder and Leafcutter}
\usage{
generate_combined_dataset(spladder_juncs, leafcutter_juncs, canonical_juncs)
}
\arguments{
\item{spladder_juncs}{A tibble the junctions identified by Spladder in
standardized format}

\item{leafcutter_juncs}{A tibble the junctions identified by Leafcutter in
standardized format}

\item{canonical_juncs}{A tibble with canonical junctions. It should contain the
column \code{junc_id} and the column \code{origin}}
}
\value{
A combined table with unique junctions. The columns RNA_tool
contains information which tools identified the given junction
}
\description{
Combines tibbles with junctions from Spladder and Leafcutter
}
\examples{
dat.combined <- generate_combined_dataset(spladder_juncs,
  leafcutter_juncs, canonical_juncs)
colnames(dat.combined)

}
