% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_junctions_RNA_tools.R
\name{generate_combined_dataset}
\alias{generate_combined_dataset}
\title{Combines tibbles with junctions from SplAdder and LeafCutter into a combined dataset of expressed splice junctions}
\usage{
generate_combined_dataset(spladder_juncs, leafcutter_juncs)
}
\arguments{
\item{spladder_juncs}{A tibble the junctions identified by SplAdder in
standardized format}

\item{leafcutter_juncs}{A tibble the junctions identified by LeafCutter in
standardized format}
}
\value{
A combined table with unique junctions. The columns RNA_tool
contains information which tools identified the given junction
}
\description{
Combines tibbles with junctions from SplAdder and LeafCutter into a combined dataset of expressed splice junctions
}
\examples{
dat.combined <- generate_combined_dataset(spladder_juncs,
  leafcutter_juncs)
colnames(dat.combined)

}
