% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmsplice.R
\name{get_exon_inclusion_junction}
\alias{get_exon_inclusion_junction}
\title{Compute the resulting junctions (junc_id) for exon inclusion of given exon
and transcript.}
\usage{
get_exon_inclusion_junction(exon_id, transcript_id, transcripts)
}
\arguments{
\item{exon_id}{A vector of exon IDs}

\item{transcript_id}{A vector of transcript IDs}

\item{transcripts}{a GRangesList with transcripts defined as GRanges of exons
created by \code{GenomicFeatures::exonsBy(txdb, by = c("tx"), use.names = TRUE)}.}
}
\value{
A list of junction IDs formatted as \verb{<chr>_<left>_<right>_<strand>}.
}
\description{
Compute the resulting junctions (junc_id) for exon inclusion of given exon
and transcript.
}
\examples{

transcript_id <- c("ENST00000243347", "ENST00000460812")
exon_id <- c("ENSE00000840477", "ENSE00003481758")
transcripts <- toy_transcripts

get_exon_inclusion_junction(exon_id, transcript_id, transcripts)

}
