% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_junc_pos.R
\name{get_junc_pos}
\alias{get_junc_pos}
\title{adds junction position in transcripts}
\usage{
get_junc_pos(tx, jx)
}
\arguments{
\item{tx}{transcripts a \code{\link[GenomicRanges]{GRangesList}} with
transcripts defined as GRanges of exons.}

\item{jx}{splice junctions as GRanges objects}
}
\value{
an integer vector with junction positions
}
\description{
adds junction position in transcripts
}
