% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_junctions_RNA_tools.R
\name{import_spladder}
\alias{import_spladder}
\title{Imports SPLADDER output from a given path}
\usage{
import_spladder(path)
}
\arguments{
\item{path}{The path to a folder with spladder output}
}
\value{
A list with tibbles. Each tibble is a Spladder output for "A5SS",
"A3SS", "cassette_exon", "intron_retention", "mutex_exons".
}
\description{
Imports SPLADDER output from a given path
}
