% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junc_to_gr.R
\name{junc_to_gr}
\alias{junc_to_gr}
\title{Convert a splice junction ID into a GRanges object}
\usage{
junc_to_gr(junc)
}
\arguments{
\item{junc}{A character vector with a junction IDs}
}
\value{
A \code{\link[GenomicRanges]{GRanges}} object
}
\description{
Convert a splice junction ID into a GRanges object
}
\examples{

}
