% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junc_to_peptide.R
\name{junc_to_peptide}
\alias{junc_to_peptide}
\title{Annotate splice junctions with resulting CDS and peptide sequence}
\usage{
junc_to_peptide(junc_id, cds, size = 30, bsg)
}
\arguments{
\item{junc_id}{junction id consisting of genomic coordinates}

\item{cds}{\code{\link[GenomicRanges]{GRangesList}} of CDS}

\item{size}{the total size of the output sequence (might be shorter if peptide is shorter)}

\item{bsg}{\code{\link[BSgenome]{BSgenome}} object such as
\code{\link[BSgenome.Hsapiens.UCSC.hg19]{BSgenome.Hsapiens.UCSC.hg19}}}
}
\value{
A data.frame with annotated input junctions
}
\description{
Annotate splice junctions with resulting CDS and peptide sequence
}
\examples{

requireNamespace("BSgenome.Hsapiens.UCSC.hg19", quietly = TRUE)
bsg <- BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19

junc_to_peptide(toy_junc_id, toy_cds, size = 30, bsg = bsg)


}
