% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafcutter_parsing.R
\name{leafcutter_transform}
\alias{leafcutter_transform}
\title{Imports "*perind.counts.gz" and "*bam.junc" from LeafCutter output and transforms the raw output
into standardized junction output format. Results of one patient should be stored in the given path.}
\usage{
leafcutter_transform(path)
}
\arguments{
\item{path}{The path to leafcutter output}
}
\value{
A tibble in standardized junction format
}
\description{
Imports "*perind.counts.gz" and "*bam.junc" from LeafCutter output and transforms the raw output
into standardized junction output format. Results of one patient should be stored in the given path.
}
