% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafcutter_parsing.R
\name{leafcutter_transform_format}
\alias{leafcutter_transform_format}
\title{Transforms LeafCutter intermediate files into standardized format}
\usage{
leafcutter_transform_format(tib)
}
\arguments{
\item{tib}{a tibble from LeafCutter intermediate format}
}
\value{
A tibble in standardized junction format
}
\description{
Transforms LeafCutter intermediate files into standardized format
}
