% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyquant.R
\name{map_requant}
\alias{map_requant}
\title{Map the re-quantification result on the junction-transcript centric data by hash id.}
\usage{
map_requant(path_to_easyquant_folder, junc_tib)
}
\arguments{
\item{path_to_easyquant_folder}{The path to easyquant folder}

\item{junc_tib}{The junction-transcript centric data predicted from WES data. A unique context sequence needs to be described by a hash id in a column \code{hash_id}}
}
\value{
Extended junction tibble with re-quantification results.  The ollowing columns are added:
\itemize{
\item \code{pos}: position of interest relative to input sequence
\item \code{junc}: reads overlapping the position of interest
\item \code{span}: read pairs spanning the position of interest
\item \code{anch}: maximal number of bases next to position of interest that are overlaped by a single read
\item \code{a}: reads mapping to sequence left of the position of interest
\item \code{b}: reads mapping to sequence right of the position of interest
}
}
\description{
Map the re-quantification result on the junction-transcript centric data by hash id.
}
\examples{


}
