% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyquant.R
\name{map_requant}
\alias{map_requant}
\title{Maps the re-quantification result from EasyQuant. on the junction-transcript centric tibble by hash id.}
\usage{
map_requant(path_to_easyquant_folder, junc_tib)
}
\arguments{
\item{path_to_easyquant_folder}{The path to EasyQuant folder}

\item{junc_tib}{The junction-transcript centric tibble, i.e. each row represents an altered transcript. Must contain a column \code{hash_id} with hash ids that relate to the column \code{name} in the Easyquant output.}
}
\value{
Extended junction-transcript tibble with re-quantification results by Easyquant. More details about Easyquant can be obtained at https://github.com/TRON-Bioinformatics/easyquant
The following columns are added:
\itemize{
\item \code{junc_interval_start}: Junction reads overlapping the junction of interest. In case of intron retentions, junctions reads that overlap the start of the intron retention interval.
\item \code{junc_interval_end}: In case of intron retentions, junctions reads that overlap the end of the intron retention interval. Is NA for non intron retention events.
\item \code{span_interval_start}: Spanning pairs overlapping the junction of interest. In case of intron retentions, spanning pairs that overlap the start of the intron retention interval.
\item \code{span_interval_end}: In case of intron retentions, spanning pairs that overlap the end of the intron retention interval. Is NA for non intron retention events.
\item \code{within_interval}: Number of reads that map into the intron retention interval. Is NA for non intron retention events.
\item \code{coverage_perc}: Intron retention interval coverage defined as the percentage of the given interval that is covered by reads. Is NA for non intron retention events.
\item \code{coverage_mean}: Intron retention interval coverage defined as the mean number of reads covering a position in the given interval. Is NA for non intron retention events.
}
}
\description{
Maps the re-quantification result from EasyQuant. on the junction-transcript centric tibble by hash id.
}
\examples{


}
