% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyquant.R
\name{read_requant}
\alias{read_requant}
\title{Imports the re-quantification results from analysis with EasyQuant}
\usage{
read_requant(path_folder)
}
\arguments{
\item{path_folder}{The path to the EasyQuant folder}
}
\value{
A tibble with with the re-quantification results. More details about Easyquant can be obtained at https://github.com/TRON-Bioinformatics/easyquant
}
\description{
Imports the re-quantification results from analysis with EasyQuant
}
