% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junc_to_peptide.R
\name{seq_extract_nonstop}
\alias{seq_extract_nonstop}
\title{Extract sequence at position from after previous stop codon and before next.}
\usage{
seq_extract_nonstop(seq, pos)
}
\arguments{
\item{seq}{Sequence}

\item{pos}{position relative to sequence}
}
\value{
a data.frame
}
\description{
Extract sequence at position from after previous stop codon and before next.
}
\examples{

seq <- "QIP*LGSNSLLFPYQLMAGSTRP*SWALGC"
seq <- c(
 "QIP*LGSNSLLFPYQLMAGSTRP*SWALGC",
 "LKMRGDTNDILSHLD*REQRVGQ*AEAASP"
)
pos <- c(14, 14)
splice2neo:::seq_extract_nonstop(seq, pos)

}
