% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spladder_parsing.R
\name{spladder_transform_exon_skipping}
\alias{spladder_transform_exon_skipping}
\title{Transforms events resulting from exon skipping from SplAdder output
format into standardized junction format}
\usage{
spladder_transform_exon_skipping(tib)
}
\arguments{
\item{tib}{A tibble in SplAdder output format}
}
\value{
A tibble in standardized junction format
}
\description{
Transforms events resulting from exon skipping from SplAdder output
format into standardized junction format
}
\examples{
spladder_output.exonskip
transformed_spladder.exonskip <- spladder_transform_exon_skipping(
 spladder_output.exonskip)
transformed_spladder.exonskip

}
