% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_stringtie.R
\name{stringtie_transform}
\alias{stringtie_transform}
\title{Imports StringTie assembled transcripts and transforms the raw output
into standardized junction output format}
\usage{
stringtie_transform(gtf.file)
}
\arguments{
\item{path}{The path to StringTie GTF file}
}
\value{
A tibble in standardized junction format
}
\description{
Imports StringTie assembled transcripts and transforms the raw output
into standardized junction output format
}
