% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_suppa2.R
\name{suppa_transform}
\alias{suppa_transform}
\title{Imports SUPPA2 output from a given path and transforms it into standardized
junction format. Results from one patients should be stored per folder. Please note
that alternative first exons (AF) and alternative last exons (AL) are not supported.}
\usage{
suppa_transform(path)
}
\arguments{
\item{path}{The path to a folder with SUPPA2 output}
}
\value{
A tibble in standardized junction format, combining all alternative
splicing classes that are covered by SUPPA2
}
\description{
Imports SUPPA2 output from a given path and transforms it into standardized
junction format. Results from one patients should be stored per folder. Please note
that alternative first exons (AF) and alternative last exons (AL) are not supported.
}
\examples{
path <-  system.file("extdata", "", package = "splice2neo")
spladder_juncs <- spladder_transform(path)
spladder_juncs

}
