% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyquant.R
\name{transform_for_requant}
\alias{transform_for_requant}
\title{Creates a table with context sequences for re-quantification analysis with EasyQuant.}
\usage{
transform_for_requant(df)
}
\arguments{
\item{df}{A data.frame with splice junctions in rows and at least the columns:
\itemize{
\item \code{cts_id} hash id for a context sequence
\item \code{cts_seq} The transcript context sequence.
\item \code{cts_junc_pos} The position of the junction within the context sequence
}}
}
\value{
A tibble that can be used for re-quantification with EasyQuant This tibble has the columns
\itemize{
\item \code{name}: Name of the input sequence.
\item \code{sequence}: The context sequence that will be requantified.
\item \code{position}: The position of the junction within the context sequence.
}
}
\description{
Creates a table with context sequences for re-quantification analysis with EasyQuant.
}
