% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafcutter_parsing.R
\name{transform_leafcutter_counts}
\alias{transform_leafcutter_counts}
\title{Transforms LeafCutter counts file into standardized junction format.}
\usage{
transform_leafcutter_counts(tib)
}
\arguments{
\item{tib}{LeafCutter counts file as tibble}
}
\value{
A tibble with columns "chromosome", "junction_start", "junction_end",
"cluster"
}
\description{
Transforms LeafCutter counts file into standardized junction format.
}
